+++ e2e/node/basic/mitm.test.ts
@@ -12,6 +12,7 @@ mitmTest(
     const counter = await createActor('tnnnb-2yaaa-aaaab-qaiiq-cai', {
       agent: await makeAgent({
         host: 'http://127.0.0.1:8888',
+        verifyQuerySignatures: false,
       }),
     });
     await expect(counter.greet('counter')).rejects.toThrow(/Invalid certificate/);